Created test base class for multi related tests
